Fix Kosinski compression of one-byte files
The main compression loop of the Kosinski compressor was entered when
the file being compressed had a size of one byte, even though it
shouldn't be. This caused the compressor to read an additional byte off
the end of the file, which was also included in the compressed stream,
such that decompressing the data would yield a file with an extra byte
at the end.