Fix a compression bug
When the compressor would try to compress a sequence of bytes that
starts with one or more zeroes followed by one or more non-zero bytes,
the compressor would mistakenly find a match starting from the virtual
zero bytes added before the start of a file and continuing into the
file. However, those virtual zero bytes can only be used for zero fills:
the decompressor will always emit zeroes when the source of a
backreference is "after" the destination.